Split FakeConsensusStateEventHandler and moved state initialization to MockConsensusStateInitializer
